requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9        # required for File / Form uploads in FastAPI

# ── Database ────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.29
asyncpg>=0.29.0                # async PostgreSQL driver
alembic>=1.13.1                # migrations (optional but recommended)

# ── Image processing ────────────────────────────────────────────────────────
Pillow>=10.3.0

# ── Storage ─────────────────────────────────────────────────────────────────
boto3>=1.34.0                  # AWS S3
cloudinary>=1.40.0             # Cloudinary

# ── Config ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1

# ── Validation ──────────────────────────────────────────────────────────────
pydantic>=2.7.0
pydantic-settings>=2.2.1
